Move is_valid_resource_id tests into existing test file
The duplicate module name (test_resource_base.py in two directories)
caused a pytest collection error. Merged tests into the existing
tests/unit/data_models/test_resource_base.py and deleted the duplicate.
