+++ tests/unit/data_models/test_resource_base.py
@@ -1,8 +1,8 @@
 """
-tests/unit/test_resource_base.py
+tests/unit/data_models/test_resource_base.py
 
 Tests for ResourceBase functionality including ID generation, timestamp handling,
-and subclass behavior.
+subclass behavior, and is_valid_resource_id().
 """
 
 import time
@@ -14,6 +14,10 @@ import pytest
 from pydantic import BaseModel, Field
 
 from bluebox.data_models.resource_base import ResourceBase
+
+
+# A valid v4 UUID to reuse across tests
+VALID_V4_UUID = "a1b2c3d4-e5f6-4a7b-8c9d-0e1f2a3b4c5d"
 
 
 class SampleResource(ResourceBase):
@@ -297,3 +301,135 @@ class TestResourceBase:
 
         assert "SampleResource" in str_repr
         assert "test_resource" in str_repr
+
+
+# --- Subclasses for is_valid_resource_id tests ---
+
+class Widget(ResourceBase):
+    pass
+
+
+class AnotherWidget(ResourceBase):
+    pass
+
+
+class TestIsValidResourceId:
+    """Tests for is_valid_resource_id()."""
+
+    # --- Happy path ---
+
+    def test_valid_id_from_create_new_id(self) -> None:
+        """ID generated by create_new_id() should always pass."""
+        assert Widget.is_valid_resource_id(Widget.create_new_id()) is True
+
+    def test_valid_id_from_instance(self) -> None:
+        """Default ID on a new instance should pass."""
+        w = Widget()
+        assert Widget.is_valid_resource_id(w.id) is True
+
+    def test_valid_manually_constructed_id(self) -> None:
+        assert Widget.is_valid_resource_id(f"Widget_{VALID_V4_UUID}") is True
+
+    # --- Wrong prefix ---
+
+    def test_wrong_class_prefix(self) -> None:
+        """Widget ID should not validate as AnotherWidget."""
+        widget_id = Widget.create_new_id()
+        assert AnotherWidget.is_valid_resource_id(widget_id) is False
+
+    def test_lowercase_prefix(self) -> None:
+        assert Widget.is_valid_resource_id(f"widget_{VALID_V4_UUID}") is False
+
+    def test_empty_prefix(self) -> None:
+        assert Widget.is_valid_resource_id(f"_{VALID_V4_UUID}") is False
+
+    def test_no_underscore_separator(self) -> None:
+        assert Widget.is_valid_resource_id(f"Widget{VALID_V4_UUID}") is False
+
+    def test_extra_prefix_text(self) -> None:
+        assert Widget.is_valid_resource_id(f"SuperWidget_{VALID_V4_UUID}") is False
+
+    def test_prefix_substring(self) -> None:
+        """'Widge' is a substring of 'Widget' but not the right prefix."""
+        assert Widget.is_valid_resource_id(f"Widge_{VALID_V4_UUID}") is False
+
+    # --- Invalid UUID portion ---
+
+    def test_empty_uuid(self) -> None:
+        assert Widget.is_valid_resource_id("Widget_") is False
+
+    def test_garbage_uuid(self) -> None:
+        assert Widget.is_valid_resource_id("Widget_not-a-uuid") is False
+
+    def test_truncated_uuid(self) -> None:
+        assert Widget.is_valid_resource_id("Widget_a1b2c3d4-e5f6-4a7b-8c9d") is False
+
+    def test_uuid_with_extra_chars(self) -> None:
+        assert Widget.is_valid_resource_id(f"Widget_{VALID_V4_UUID}-extra") is False
+
+    def test_uuid_with_braces(self) -> None:
+        """UUIDs wrapped in braces are valid for stdlib UUID(), but the format should still parse."""
+        braced = f"Widget_{{{VALID_V4_UUID}}}"
+        assert Widget.is_valid_resource_id(braced) is True
+
+    def test_uuid_v1_rejected(self) -> None:
+        """A valid UUID v1 should be rejected (not v4)."""
+        v1_uuid = "a1b2c3d4-e5f6-1a7b-8c9d-0e1f2a3b4c5d"
+        assert Widget.is_valid_resource_id(f"Widget_{v1_uuid}") is False
+
+    def test_uuid_v3_rejected(self) -> None:
+        v3_uuid = "a1b2c3d4-e5f6-3a7b-8c9d-0e1f2a3b4c5d"
+        assert Widget.is_valid_resource_id(f"Widget_{v3_uuid}") is False
+
+    def test_uuid_v5_rejected(self) -> None:
+        v5_uuid = "a1b2c3d4-e5f6-5a7b-8c9d-0e1f2a3b4c5d"
+        assert Widget.is_valid_resource_id(f"Widget_{v5_uuid}") is False
+
+    # --- Edge-case inputs ---
+
+    def test_empty_string(self) -> None:
+        assert Widget.is_valid_resource_id("") is False
+
+    def test_just_classname(self) -> None:
+        assert Widget.is_valid_resource_id("Widget") is False
+
+    def test_just_underscore(self) -> None:
+        assert Widget.is_valid_resource_id("_") is False
+
+    def test_multiple_underscores_in_uuid(self) -> None:
+        """Underscores in the UUID portion should fail."""
+        assert Widget.is_valid_resource_id("Widget_a1b2c3d4_e5f6_4a7b_8c9d_0e1f2a3b4c5d") is False
+
+    def test_whitespace_around_id(self) -> None:
+        assert Widget.is_valid_resource_id(f" Widget_{VALID_V4_UUID} ") is False
+
+    def test_newline_in_id(self) -> None:
+        assert Widget.is_valid_resource_id(f"Widget_{VALID_V4_UUID}\n") is False
+
+    def test_double_underscore_separator(self) -> None:
+        assert Widget.is_valid_resource_id(f"Widget__{VALID_V4_UUID}") is False
+
+    # --- Called on ResourceBase directly ---
+
+    def test_raises_on_resource_base(self) -> None:
+        with pytest.raises(TypeError, match="is_valid_resource_id.*must be called on a subclass"):
+            ResourceBase.is_valid_resource_id(f"ResourceBase_{VALID_V4_UUID}")
+
+    # --- Cross-class isolation ---
+
+    def test_each_subclass_validates_its_own_prefix(self) -> None:
+        """Two subclasses should not accept each other's IDs."""
+        widget_id = Widget.create_new_id()
+        another_id = AnotherWidget.create_new_id()
+
+        assert Widget.is_valid_resource_id(widget_id) is True
+        assert Widget.is_valid_resource_id(another_id) is False
+        assert AnotherWidget.is_valid_resource_id(another_id) is True
+        assert AnotherWidget.is_valid_resource_id(widget_id) is False
+
+    # --- Bulk / fuzz-lite ---
+
+    def test_many_generated_ids_all_valid(self) -> None:
+        """100 generated IDs should all validate."""
+        for _ in range(100):
+            assert Widget.is_valid_resource_id(Widget.create_new_id()) is True
